cat_dog_requirements.txt
streamlit
tensorflow
numpy
pillow
scikit-learn
gdown